satisfy codacy with a net positive change
