+++ dockstore-client/src/main/java/io/dockstore/client/cli/nested/notificationsclients/NotificationsClient.java
@@ -13,7 +13,7 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-package io.dockstore.client.cli.nested.NotificationsClients;
+package io.dockstore.client.cli.nested.notificationsclients;
 
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
